+++ js/task-screen.js
@@ -2,11 +2,12 @@ import {render, changeScreen} from './util.js';
 import introElement from './intro.js';
 // import statsElement from './stats.js';
 import state from './game-data.js';
-import tasks from './game-data.js';
+import tasks from './task-file.js';
 import gameHeaderTemplate from './game-header.js';
 import gameOneElement from './game-1.js';
 import gameTwoElement from './game-2.js';
 import gameThreeElement from './game-3.js';
+
 
 const gameTask = (currentState, task) => {
   let template = ``;
@@ -27,11 +28,17 @@ const gameTask = (currentState, task) => {
   return template;
 };
 
+
 const getCurrentLevel = (currentState, task) => {
   return render(gameHeaderTemplate(currentState) + gameTask(currentState, task));
 };
 
+let levelScreen = getCurrentLevel(state, tasks[0]);
+
 const levelHandler = (currentLevel, task, nextLevel) => {
+  if (state.level < 9) {
+    state.level++;
+  }
   switch (task.number) {
     case 2: {
       const gameOptions = Array.prototype.slice.call(currentLevel.querySelectorAll(`.game__answer input[type="radio"]`));
@@ -68,14 +75,12 @@ const levelHandler = (currentLevel, task, nextLevel) => {
     }
   }
 };
-const levelScreen = getCurrentLevel(state, tasks[0]);
-
 const backBtn = levelScreen.querySelector(`.back`);
 backBtn.addEventListener(`click`, () => {
   changeScreen(introElement);
 });
 
-levelHandler(levelScreen, tasks[0], getCurrentLevel(state, tasks[1]));
+levelHandler(levelScreen, tasks[state.level], getCurrentLevel(state, tasks[state.level + 1]));
 export default levelScreen;
 
 // container.appendChild(render(gameTask(state, tasks[0])));
